Update updateTabProperties Internal + API
