+++ src/demo/tabapp2.ts
@@ -9,7 +9,7 @@ document.addEventListener('DOMContentLoaded', () => {
 });
 
 fin.desktop.main(() => {
-    Layouts.setTabClient('http://localhost:1337/demo/tabstrips/custom2.html', {height: 62});
+    Layouts.setTabstrip({url: 'http://localhost:1337/demo/tabstrips/custom2.html', height: 62});
     Layouts.addEventListener('join-tab-group', () => {
         console.log('TABBED: ');
     });
